Change compound OnMouseOver handling
This sends Notify events to the compound window, if the mouse
enters to any of the childrens, or leaves from any of the children
outside of the compound window / control.

Additionally all leave events from the compound window are
dropped, if the mouse moves to a child control.

This allows us to skip a lot of redraws AKA flickers for border
focus indicating themes (mbNoFocusRects).

Change-Id: Ibc03226b1cb86443b62a688d3b98f17811b3db61