Rename POST messages to DEFERRED
POST indicates a message is processed after the original message.
While this is semantically correct, these messages are used to defer
the processing of a message AKA postponed, until we acquire the
Yield mutex in the main thread.

While at it rename POSTFOCUS to DEFERREDSETFOCUS, as this defers
the WM_SETFOCUS message.

Change-Id: I66669e1a1ad807049232476414ac9e2fd4e5cb47