Simplify Windows system timer / LO event handling
The system timer is supposed to have the following constraints:

1. There is just one system timer, which drives LO event loop
2. The timer has to run in the main window thread
3. Messages are processed in order, so there is no real need for
   calling SendMessageW instead of PostMessageW
4. LO event loop is supposed to be processed in the main window
   thread with the Solar mutex acquired
5. The system timer is a single-shot timer

In the end this removes duplicated code and simplifies it.

Drop the 16bit MAX duration limit, as CreateTimerQueueTimer uses
a DWORD for the DueTime parameter, which is always UINT32.

CreateTimerQueueTimer already sets the period to 0, which makes it
a one-shot timer, but there is also the WT_EXECUTEONLYONCE, which
enforces the Period parameter to be 0.

Change-Id: I549142394334bdc098f053b42f222b23cf4fcecd